Added the ability to export vectors as multiple PNG mask files.
Created an image classification object which holds all vectors, and data about the image being classified.

Added a function to the Vector class which allows for drawing of the PNG mask of a Vector, given a graphics object.

Added a status bar, moved the picturebox into its own panel with autoscroll enabled, so that the menubar and statusbar will always be visible and so the picturebox will be scrollable.

Added reference to Nuget package Ookii, for its Vista folder browser dialog.

Added ability to zoom in and out of the image being classified.

TODO: Get zoom logic to work with point logic (if you're zoomed in, get it to get the ACTUAL coordinate you're clicking, and get it to paint the coordinate at the correct position when zoomed in)